+++ src/effitor/plugins/mark/element.ts
@@ -34,11 +34,13 @@ export class EtMarkElement extends EtRichTextElement {
         }
     }
     focusoutCallback(ctx: Et.EditorContext): void {
+        // 已不在dom上, 跳过
+        if (!this.isConnected) return
         // 若标记节点仅剩零宽字符，应当将其移除
         if (this.textContent === HtmlCharEnum.ZERO_WIDTH_SPACE) {
             // 回调在selchange之后触发，光标已不在this节点; 标记节点内光标位置，以便撤回删除时能让光标落于this节点内
             const srcTr = dom.caretStaticRangeInNode(this, this.childNodes.length)
-            removeNodeAndMerge(ctx, this, false, srcTr) && ctx.commandHandler.handle()
+            removeNodeAndMerge(ctx, this, srcTr, false) && ctx.commandHandler.handle()
         }
         else {
             this.classList.remove(MarkStatus.HINTING)